backend/requirements.txt
fastapi
uvicorn
librosa
numpy
openai-whisper
pyphen
pydantic
scikit-learn
python-multipart
textstat
numba
